requirements.txt
Django>=4.2
djangorestframework
pymongo
python-dotenv
requests
qrcode
Pillow
django-cors-headers